Extract interrupt and reasoning flows
