+++ src/tests/helpers/readInteractiveAppSource.ts
@@ -5,7 +5,9 @@ const INTERACTIVE_APP_SOURCE_FILES = [
   "../../cli/app/useSubmitHandler.ts",
   "../../cli/app/useConversationLoop.ts",
   "../../cli/app/useConfigurationHandlers.ts",
+  "../../cli/app/useInterruptHandler.ts",
   "../../cli/app/AppCoordinator.tsx",
+  "../../cli/app/useReasoningCycle.ts",
   "../../cli/app/useApprovalFlow.ts",
   "../../cli/app/useConversationSwitching.ts",
   "../../cli/app/useBashHandlers.ts",
